+++ my-app/src/App.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import Header from './components/Header';
-import Home from './components/Home';
+import Home from './components/Home/Home';
 import Donate from './components/Donate';
 import Members from './components/Members';
 import Archive from './components/Archive';
